+++ src/ast/nodes/CatchClause.ts
@@ -8,26 +8,27 @@ import { PatternNode } from './shared/Pattern';
 
 export default class CatchClause extends NodeBase {
 	type: NodeType.tCatchClause;
-	param: PatternNode;
+	param: PatternNode | null;
 	body: BlockStatement;
 
 	scope: CatchScope;
 	preventChildBlockScope: true;
 
 	createScope(parentScope: Scope) {
-		this.scope = new CatchScope(parentScope);
+		this.scope = new CatchScope(parentScope, this.context.deoptimizationTracker);
 	}
 
 	initialise() {
 		this.included = false;
-		this.param.declare('parameter', UNKNOWN_EXPRESSION);
+
+		if (this.param) {
+			this.param.declare('parameter', UNKNOWN_EXPRESSION);
+		}
 	}
 
 	parseNode(esTreeNode: GenericEsTreeNode) {
-		this.body = <BlockStatement>new this.context.nodeConstructors.BlockStatement(
-			esTreeNode.body,
-			this,
-			this.scope
+		this.body = <BlockStatement>(
+			new this.context.nodeConstructors.BlockStatement(esTreeNode.body, this, this.scope)
 		);
 		super.parseNode(esTreeNode);
 	}