+++ test/function/samples/deconflict-default-exported-named-function/named.js
@@ -1,0 +1,3 @@
+export const test = function() {
+	return 'named';
+};